Move shell inside commands.
