+++ hatch/commands/__init__.py
@@ -14,3 +14,4 @@ from hatch.commands.release import release
 from hatch.commands.python_path import python_path
 from hatch.commands.env import env
 from hatch.commands.shed import shed
+from hatch.commands.shell import shell
